+++ features/fixtures/maze_runner/Assets/Scripts/MobileScenarioRunner.cs
@@ -35,6 +35,10 @@ public class MobileScenarioRunner : MonoBehaviour {
         {"18", "Android Persistence Directory" },
         {"19", "Disabled Release Stage" },
         {"20", "Enabled Release Stage" },
+        {"21", "Java Background Crash No Threads" },
+        {"22", "iOS Native Error" },
+        {"23", "iOS Native Error No Threads" },
+
 
 
 
@@ -137,6 +141,10 @@ public class MobileScenarioRunner : MonoBehaviour {
             case "Native exception":
                 config.ProjectPackages = new string[] { "test.test.test" };
                 break;
+            case "iOS Native Error No Threads":
+            case "Java Background Crash No Threads":
+                config.SendThreads = ThreadSendPolicy.NEVER;
+                break;
             case "Start SDK no errors":
             case "Disable Native Errors":
                 config.EnabledErrorTypes = new ErrorTypes[0];
@@ -194,9 +202,9 @@ public class MobileScenarioRunner : MonoBehaviour {
             case "Start SDK":
             case "Start SDK no errors":
                 break;
+            case "iOS Native Error":
+            case "iOS Native Error No Threads":
             case "Discard Error Class":
-                NativeException();
-                break;
             case "Disable Native Errors":
                 NativeException();
                 break;
@@ -213,6 +221,7 @@ public class MobileScenarioRunner : MonoBehaviour {
                 SetUser();
                 LogError();
                 break;
+            case "Java Background Crash No Threads":
             case "Java Background Crash":
                 AddMetadataForRedaction();
                 MobileNative.TriggerBackgroundJavaCrash();